feat(controller:task): Move tasks_controller to controllers/api/v1 feat(controller:home): Add home_controller feat(model:task): Validate for due_date feat(model:user): Add has_many :tasks feat(routes): Add name_spaces "api" and "v1"
