+++ app/controllers/home_controller.rb
@@ -1,0 +1,4 @@
+class HomeController < ActionController
+  def index
+  end
+end
