+++ app/models/task.rb
@@ -1,3 +1,5 @@
 class Task < ApplicationRecord
   belongs_to :user
+
+  validates :due_date, presence: true
 end
