+++ app/models/user.rb
@@ -1,3 +1,5 @@
 class User < ApplicationRecord
   has_secure_password
+
+  has_many :tasks
 end
