+++ config/routes.rb
@@ -1,4 +1,12 @@
 Rails.application.routes.draw do
-  resources :tasks
+
+  namespace :api do
+    namespace :v1 do
+      resources :tasks
+    end
+  end
+
+  root to: "home#index"
+
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 end
